Separate Auth and User controller
