+++ app/routes/auth.routes.js
@@ -1,13 +1,13 @@
 const express = require("express");
 const router = express.Router();
-const userController = require("../controllers/user.controller");
+const authController = require("../controllers/auth.controller");
 const authMiddleware = require("../middleware/auth.middleware");
 
-router.post("/signup", userController.signup);
-router.post("/login", userController.login);
-router.get("/read", authMiddleware, userController.readMe);
-router.get("/export", authMiddleware, userController.exportMe);
-router.put("/update", authMiddleware, userController.update);
-router.delete("/delete", authMiddleware, userController.delete)
+router.post("/signup", authController.signup);
+router.post("/login", authController.login);
+router.get("/read", authMiddleware, authController.readMe);
+router.get("/export", authMiddleware, authController.exportMe);
+router.put("/update", authMiddleware, authController.update);
+router.delete("/delete", authMiddleware, authController.delete)
 
 module.exports = router;
